Move rpc method into the correct file
