+++ Assets/Scripts/Multiplayer/MultiplayerAudioSourcesManager.cs
@@ -13,8 +13,6 @@
 // limitations under the License.
 
 using System.Collections.Generic;
-using Fusion;
-using Photon.Voice.Unity;
 using TiltBrush;
 using UnityEngine;
 
@@ -106,19 +104,6 @@ namespace OpenBrush
             if (sources.TryGetValue(playerId, out AudioSource source))
             {
                 source.mute = value;
-            }
-        }
-
-        [Rpc(RpcSources.All, RpcTargets.All)]
-        public static void RPC_MutePlayer(NetworkRunner runner, bool mute, PlayerRef targetPlayer)
-        {
-            if (runner.TryGetPlayerObject(targetPlayer, out NetworkObject networkObject))
-            {
-                var speaker = networkObject.GetComponent<Speaker>(); // Photon Voice 2 component
-                if (speaker != null)
-                {
-                    speaker.enabled = !mute; // Disables speaker to mute
-                }
             }
         }
     }
