+++ Assets/Scripts/Multiplayer/Photon/PhotonRPC.cs
@@ -23,6 +23,7 @@ using Fusion;
 using TiltBrush;
 using static TiltBrush.SketchControlsScript;
 using System.Threading.Tasks;
+using Photon.Voice.Unity;
 
 namespace OpenBrush.Multiplayer
 {
@@ -593,6 +594,19 @@ namespace OpenBrush.Multiplayer
             MultiplayerManager.m_Instance.Disconnect();
         }
 
+        [Rpc(RpcSources.All, RpcTargets.All)]
+        public static void RPC_MutePlayer(NetworkRunner runner, bool mute, PlayerRef targetPlayer)
+        {
+            if (runner.TryGetPlayerObject(targetPlayer, out NetworkObject networkObject))
+            {
+                var speaker = networkObject.GetComponent<Speaker>(); // Photon Voice 2 component
+                if (speaker != null)
+                {
+                    speaker.enabled = !mute; // Disables speaker to mute
+                }
+            }
+        }
+        
         #endregion
     }
 }
